CPIDataBase

CPIDataBase.CPIDataBase

CPIDataBase.AbstractCPIBase

CPIDataBase.CODETYPE

CPIDataBase.CPIIndex

CPIDataBase.CPIResult

CPIDataBase.CPITree

CPIDataBase.CPIVarInterm

CPIDataBase.CombinationFunction

CPIDataBase.CountryStructure

CPIDataBase.DATETYPE

CPIDataBase.DESCTYPE

CPIDataBase.EnsembleFunction

CPIDataBase.EnsembleInflationFunction

CPIDataBase.FullCPIBase

CPIDataBase.FullCPIBase

CPIDataBase.Group

CPIDataBase.IndexCPIBase

CPIDataBase.IndexCPIBase

CPIDataBase.InflationCombination

CPIDataBase.InflationEnsemble

CPIDataBase.InflationFunction

CPIDataBase.Item

CPIDataBase.MixedCountryStructure

CPIDataBase.UniformCountryStructure

CPIDataBase.VarCPIBase

CPIDataBase.VarCPIBase

CPIDataBase.capitalize

CPIDataBase.capitalize

CPIDataBase.capitalize!

CPIDataBase.components

CPIDataBase.components

CPIDataBase.compute_index

CPIDataBase.cpi_tree_nodes

CPIDataBase.get_cpi_tree

CPIDataBase.getdates

CPIDataBase.getdates

CPIDataBase.getunionalltype

CPIDataBase.getunionalltype

CPIDataBase.infl_dates

CPIDataBase.infl_periods

CPIDataBase.measure_name

CPIDataBase.measure_name

CPIDataBase.measure_name

CPIDataBase.measure_name

CPIDataBase.measure_name

CPIDataBase.measure_name

CPIDataBase.measure_tag

CPIDataBase.num_measures

CPIDataBase.params

CPIDataBase.periods

CPIDataBase.periods

CPIDataBase.varinteran

CPIDataBase.varinteran!

CPIDataBase.varinterm

CPIDataBase.varinterm

CPIDataBase.varinterm!

CPIDataBase.weights

CPIDataBase.TestHelpers.getbasedates

CPIDataBase.TestHelpers.getrandombase

CPIDataBase.TestHelpers.getrandomcountryst

CPIDataBase.TestHelpers.getrandomweights

CPIDataBase.TestHelpers.getzerobase

CPIDataBase.TestHelpers.getzerocountryst

CPIDataBase.CPIDataBase — Module

CPIDataBase

Librería base para tipos y funcionalidad básica para manejo de los datos desagregados del IPC a nivel de república.

CPIDataBase.CODETYPE — Type

const CODETYPE = Union{Vector{String}, Nothing}

Tipos posibles para los códigos en el campo codes de un FullCPIBase.

CPIDataBase.DATETYPE — Type

const DATETYPE = Union{StepRange{Date, Month}, Vector{Date}}

Tipos posibles para el campo de fechas dates de un AbstractCPIBase.

CPIDataBase.DESCTYPE — Type

const DESCTYPE = Union{Vector{String}, Nothing}

Tipos posibles para los nombres en el campo names de un FullCPIBase.

CPIDataBase.AbstractCPIBase — Type

abstract type AbstractCPIBase{T <: AbstractFloat}

Tipo abstracto para representar conjuntos de colecciones de datos del IPC.

Vea también: FullCPIBase, VarCPIBase e IndexCPIBase.

CPIDataBase.CPIIndex — Type

Tipo concreto único para obtener el índice de una función de inflación

CPIDataBase.CPIResult — Type

Tipo abstracto para manejar el despacho de las funciones de inflación

CPIDataBase.CPITree — Type

CPITree

CPITree(base::FullCPIBase, tree::Union{Group, Item}, group_names::Vector{String}, group_codes::Vector{String})
CPITree(; base::FullCPIBase, groupsdf::DataFrame, characters::(NTuple{N, Int} where N), upperlevel_code = "_0", upperlevel_name = "IPC")

Contenedor envolvente de un árbol jerárquico del IPC y los datos necesarios de los gastos básicos para computar cualquier jerarquía dentro del árbol. Permite visualizar y explorar la composición del IPC de un país, así como computar los índices de precios de las diferentes jerarquías de la estructura del IPC. Está compuesto por:

Un objeto base, de tipo FullCPIBase, el cual almacena las series de tiempo de los índices de los gastos básicos.

Un objeto tree que contiene la estructura de nodos del IPC.

El vector de nombres group_names de los grupos del árbol tree.

El vector de códigos group_codes de los grupos del árbol tree.

El constructor simple requiere una estructura jerárquica de nodos como la devuelta por get_cpi_tree. Al utilizar el constructor con groupsdf y characters, se construye automáticamente un CPITree utilizando los códigos como indicadores de la estructura jerárquica. Los códigos de los gastos básicos contenidos en base describen cómo se agrupan las jerarquías. Por ejemplo, el siguiente FullCPIBase contiene 10 gastos básicos:

julia> base
FullCPIBase{Float32, Float32}: 36 períodos × 10 gastos básicos Jan-01-Dec-03
┌─────┬────────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
│ Row │      Dates │ _011101 │ _011201 │ _011202 │ _021101 │ _022101 │ _031101 │ _041101 │ _041201 │ _041202 │ _041301 │
│     │            │  Item A │  Item B │  Item C │  Item D │  Item E │  Item F │  Item G │  Item H │  Item I │  Item J │
│     │            │ 11.8947 │ 2.39931 │ 7.80836 │ 1.58585 │  20.141 │ 12.0526 │ 9.56901 │ 14.6256 │ 15.0276 │ 4.89602 │
├─────┼────────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│   1 │ 2001-01-01 │  100.00 │  100.00 │  100.00 │  100.00 │  100.00 │  101.13 │  100.00 │  100.00 │  100.51 │  100.00 │
│   2 │ 2001-02-01 │  100.00 │  100.00 │  100.00 │  100.00 │  100.00 │  102.28 │  100.00 │  100.00 │  101.02 │  100.00 │
│   3 │ 2001-03-01 │  100.00 │  100.00 │  100.00 │  100.00 │  100.00 │  103.44 │  100.00 │  100.00 │  101.53 │  100.00 │
│   4 │ 2001-04-01 │  100.00 │  100.00 │  100.00 │  100.00 │  100.00 │  104.61 │  100.00 │  100.00 │  102.05 │  100.00 │
│   5 │ 2001-05-01 │  100.00 │  100.00 │  100.00 │  100.00 │  100.00 │  105.79 │  100.00 │  100.00 │  102.56 │  100.00 │
│  ⋮  │     ⋮      │    ⋮    │    ⋮    │    ⋮    │    ⋮    │    ⋮    │    ⋮    │    ⋮    │    ⋮    │    ⋮    │    ⋮    │
│  32 │ 2003-08-01 │  100.00 │  100.00 │  100.00 │  100.00 │  100.00 │  143.39 │  100.00 │  100.00 │  117.59 │  100.00 │
│  33 │ 2003-09-01 │  100.00 │  100.00 │  100.00 │  100.00 │  100.00 │  145.02 │  100.00 │  100.00 │  118.19 │  100.00 │
│  34 │ 2003-10-01 │  100.00 │  100.00 │  100.00 │  100.00 │  100.00 │  146.66 │  100.00 │  100.00 │  118.79 │  100.00 │
│  35 │ 2003-11-01 │  100.00 │  100.00 │  100.00 │  100.00 │  100.00 │  148.32 │  100.00 │  100.00 │  119.39 │  100.00 │
│  36 │ 2003-12-01 │  100.00 │  100.00 │  100.00 │  100.00 │  100.00 │  150.00 │  100.00 │  100.00 │  120.00 │  100.00 │
└─────┴────────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘
                                                                                                         26 rows omitted

En este ejemplo, el argumento debe especificarse como characters = (3, 4, 5, 7), pues los códigos indican las jerarquías en el IPC de manera siguiente:

Los primeros 3 caracteres indican la jerarquía de división de gasto.

El siguiente caracter indica la jerarquía de agrupación de gasto.

El siguiente caracter indica la jerarquía de subgrupo de gasto.

Los siguientes 2 caracteres indican el número de gasto básico dentro de su grupo.

Por su parte, el DataFrame groupsdf debe tener la estructura mínima siguiente:

La primera columna debe ser de tipo String y contiene los códigos de los grupos disponibles en la estructura del IPC.

La segunda columna debe ser de tipo String y contiene las descripciones o

nombres de los grupos disponibles en la estructura del IPC. Por ejemplo, el DataFrame groupsdf puede verse de esta forma:

17×2 DataFrame
 Row │ code    description 
     │ String  String      
─────┼─────────────────────
   1 │ _01     Div._01
   2 │ _02     Div._02
   3 │ _03     Div._03
   4 │ _04     Div._04
   5 │ _011    Agr._011
   6 │ _021    Agr._021
   7 │ _022    Agr._022
   8 │ _031    Agr._031
   9 │ _041    Agr._041
  10 │ _0111   Subgr._0111
  11 │ _0112   Subgr._0112
  12 │ _0211   Subgr._0211
  13 │ _0221   Subgr._0221
  14 │ _0311   Subgr._0311
  15 │ _0411   Subgr._0411
  16 │ _0412   Subgr._0412
  17 │ _0413   Subgr._0413

Por ejemplo, al construir un CPITree con la estructura de códigos indicada anteriormente y el DataFrame de ejemplo, el árbol del IPC se puede ver de esta forma:

julia> tree = CPITree(; base, groupsdf, characters=(3,4,5,7))
CPITree{Group{Group{Group{Group{Item{Float32}, Float32}, Float32}, Float32}, Float32}} con datos
└─→ FullCPIBase{Float32, Float32}: 36 períodos × 10 gastos básicos Jan-01-Dec-03
_0: IPC [100.0]
├─ _01: Div._01 [21.491905]
│  └─ _011: Agr._011 [21.491905]
│     ├─ _0111: Subgr._0111 [7.352945]
│     │  └─ _011101: Item A [7.352945]
│     └─ _0112: Subgr._0112 [14.13896]
│        ├─ _011201: Item B [6.7442417]
│        └─ _011202: Item C [7.394718]
├─ _02: Div._02 [3.0530455]
│  ├─ _021: Agr._021 [1.1036392]
│  │  └─ _0211: Subgr._0211 [1.1036392]
│  │     └─ _021101: Item D [1.1036392]
│  └─ _022: Agr._022 [1.9494063]
│     └─ _0221: Subgr._0221 [1.9494063]
│        └─ _022101: Item E [1.9494063]
├─ _03: Div._03 [11.68543]
│  └─ _031: Agr._031 [11.68543]
│     └─ _0311: Subgr._0311 [11.68543]
│        └─ _031101: Item F [11.68543]
└─ _04: Div._04 [63.769615]
   └─ _041: Agr._041 [63.769615]
      ├─ _0411: Subgr._0411 [16.103952]
      │  └─ _041101: Item G [16.103952]
      ├─ _0412: Subgr._0412 [28.824577]
      │  ├─ _041201: Item H [11.367162]
      │  └─ _041202: Item I [17.457417]
      └─ _0413: Subgr._0413 [18.841085]
         └─ _041301: Item J [18.841085]

CPIDataBase.CPIVarInterm — Type

Tipo concreto único para obtener la variación intermensual de una función de inflación

CPIDataBase.CombinationFunction — Type

CombinationFunction{N, W} <: EnsembleInflationFunction <: InflationFunction

CombinationFunction(ensemble, weights [, name, tag])
CombinationFunction(inflfn1, inflfn2 [, ...], weights [, name, tag])

Función de inflación para computar un promedio ponderado de un conjunto de N de medidas de inflación con tipo del vector de ponderaciones W.

CPIDataBase.CountryStructure — Type

CountryStructure{N, T <: AbstractFloat}

Tipo abstracto que representa el conjunto de bases del IPC de un país.

CPIDataBase.EnsembleFunction — Type

EnsembleFunction{N} <: EnsembleInflationFunction

EnsembleFunction(inflfn1, inflfn2 [, ...])

Función de inflación para computar un conjunto de N de medidas de inflación simultáneamente utilizando las funciones inflfn1, inflfn2, ....

CPIDataBase.EnsembleInflationFunction — Type

EnsembleInflationFunction <: InflationFunction <: InflationFunction

Tipo abstracto para conformar conjuntos de medidas de inflación

CPIDataBase.FullCPIBase — Type

FullCPIBase{T, B} <: AbstractCPIBase{T}

FullCPIBase(ipc::Matrix{T}, v::Matrix{T}, w::Vector{T}, dates::DATETYPE, baseindex::B, codes::CODETYPE, names::DESCTYPE) where {T, B}
FullCPIBase(df::DataFrame, gb::DataFrame)

Contenedor completo para datos desagregados del IPC de un país. Se representa por:

Matriz de índices de precios ipc que incluye la fila con los índices del número base.

Matriz de variaciones intermensuales v. En las filas contiene los períodos y en las columnas contiene los gastos básicos.

Vector de ponderaciones w de los gastos básicos.

Fechas correspondientes dates (por meses).

Índices base baseindex.

Códigos y nombres de los gastos básicos en codes y names.

El tipo T representa el tipo de datos para representar los valores de punto flotante. El tipo B representa el tipo del campo baseindex; por ejemplo, Float32 o Vector{Float32}.

CPIDataBase.FullCPIBase — Method

FullCPIBase(df::DataFrame, gb::DataFrame)

Este método constructor devuelve una estructura FullCPIBase a partir de los DataFrames de índices de precios df y de descripción de los gastos básicos gb.

El DataFrame df posee la siguiente estructura:

Contiene en la primera columna las fechas o períodos de los datos. En las siguientes columnas, debe contener los códigos de cada una de las categorías o gastos básicos de la estructura del IPC, junto con la serie de tiempo con los índices de precios individuales.

En las filas del DataFrame contiene los períodos por meses. La primera fila del DataFrame se utiliza para obtener el índice base. Si el valor es el mismo para todos los gastos básicos, se tomará únicamente este valor escalar (por ejemplo 100.0 como un Float64). En algunos casos, es posible que no se disponga de la información completa, por lo que los índices base podrían ser diferentes entre sí. En este caso, baseindex almacenará el vector de índices base originales.

Un ejemplo de cómo puede verse este DataFrame es el siguiente:

121×219 DataFrame
 Row │ Date         _011111  _011121  _011131  _011141  _011142  _011151  _011152 ⋯
     │ Date         Float64  Float64  Float64  Float64  Float64  Float64  Float64 ⋯
─────┼─────────────────────────────────────────────────────────────────────────────
   1 │ 2000-12-01   100.0    100.0    100.0    100.0    100.0    100.0    100.0   ⋯
   2 │ 2001-01-01   100.55   103.23   101.66   106.47   100.36   100.0    102.57   
   3 │ 2001-02-01   101.47   104.82   102.73   108.38   101.37   100.0    103.35   
   4 │ 2001-03-01   101.44   107.74   104.9    103.76   101.32   100.0    104.27   
   5 │ 2001-04-01   101.91   107.28   106.19   107.83   101.82   100.0    104.73  ⋯
   6 │ 2001-05-01   102.77   106.12   106.9    109.16   101.81   100.0    105.21   
   7 │ 2001-06-01   103.23   109.04   107.4    112.13   102.72   100.0    105.47   
   8 │ 2001-07-01   104.35   112.72   107.96   117.19   105.09   100.0    105.66   
  ⋮  │     ⋮          ⋮        ⋮        ⋮        ⋮        ⋮        ⋮        ⋮     ⋱
 114 │ 2010-05-01   218.45   501.39   200.28   477.5    179.0    215.0    164.16  ⋯
 115 │ 2010-06-01   219.28   503.35   203.88   476.26   180.94   214.02   164.97   
 116 │ 2010-07-01   219.1    503.78   205.19   478.34   181.78   217.6    165.9    
 117 │ 2010-08-01   218.52   507.45   206.87   486.72   181.51   223.76   166.46   
 118 │ 2010-09-01   218.9    505.8    206.45   501.23   182.04   228.34   166.04  ⋯
 119 │ 2010-10-01   219.51   504.41   205.78   504.4    182.35   221.98   166.3    
 120 │ 2010-11-01   219.11   509.63   205.41   502.88   182.16   217.01   166.34   
 121 │ 2010-12-01   218.79   511.38   205.09   506.04   182.14   218.63   165.99   
                                                   211 columns and 105 rows omitted

El DataFrame gb posee la siguiente estructura:

La primera columna contiene los códigos de las columnas del DataFrame df.

La segunda columna contiene el nombre o la descripción de cada una de las categorías en las columnas de df.

Y finalmente, la tercer columna, debe contener las ponderaciones asociadas a cada una de las categorías o gastos básicos de las columnas de df.

Los nombres de las columnas no son tomados en cuenta, solamente el orden y los tipos.

Un ejemplo de cómo puede verse este DataFrame es el siguiente:

218×3 DataFrame
 Row │ Code     GoodOrService                       Weight
     │ String   String                              Float64     
─────┼────────────────────────────────────────────────────────
   1 │ _011111  Arroz                               0.483952
   2 │ _011121  Pan                                 2.82638
   3 │ _011131  Pastas frescas y secas              0.341395
   4 │ _011141  Productos de tortillería            1.69133
  ⋮  │   ⋮                     ⋮                     ⋮
 216 │ _093111  Gastos por seguros                  0.236691
 217 │ _093121  Gastos por servicios funerarios     0.289885
 218 │ _094111  Gastos por servicios diversos pa…   0.151793
                                              211 rows omitted

CPIDataBase.Group — Type

Group{S,T}

Group(code, name, children::Vector{S}) where S
Group(code, name, children...)

Representa un nodo de agrupación de cualquier nivel en la estructura de nodos del IPC. Puede almacenar gastos básicos u otros grupos de mayor jerarquía. Almacena el código del grupo, su nombre o descripción y su ponderación en el IPC.

Posee los campos:

code: almacena el código del grupo como un String.

name: almacena el nombre del grupo como un String.

weight::T: almacena la ponderación del grupo como un valor flotante de tipo T.

children::Vector{S}: almacena el vector de nodos "hijos" de la estructura. Por ejemplo, este vector podría ser un vector de elementos Item para agrupar un conjunto de gastos básicos.

Aunque este nodo será usualmente creado de manera automática por métodos como CPITree, se pueden crear estructuras jerárquicas manualmente. Por ejemplo, para crear un grupo:

julia> a = Item("_011201", "Item B", 6.7442417f0)
Item{Float32}("_011201", "Item B", 6.7442417f0)

julia> b = Item("_011202", "Item C", 7.394718f0)
Item{Float32}("_011202", "Item C", 7.394718f0)

julia> g = Group("_0112", "Subgr._0112", a, b)
Group{Item{Float32}, Float32}
_0112: Subgr._0112 [14.13896] 
├─ _011201: Item B [6.7442417]
└─ _011202: Item C [7.394718]

Ver también: Item, CPITree.

CPIDataBase.IndexCPIBase — Type

IndexCPIBase{T, B} <: AbstractCPIBase{T}

IndexCPIBase(ipc::Matrix{T}, w::Vector{T}, dates::DATETYPE, baseindex::B) where {T, B}

Contenedor genérico de índices de precios del IPC de un país. Se representa por:

Matriz de índices de precios ipc que incluye la fila con los índices del númbero base.

Vector de ponderaciones w de los gastos básicos.

Fechas correspondientes dates (por meses).

Índices base baseindex.

El tipo T representa el tipo de datos para representar los valores de punto flotante. El tipo B representa el tipo del campo baseindex; por ejemplo, Float32 o Vector{Float32}.

CPIDataBase.IndexCPIBase — Method

IndexCPIBase(df::DataFrame, gb::DataFrame)

Este constructor devuelve una estructura IndexCPIBase a partir del DataFrame de índices de precios df, que contiene en las columnas las categorías o gastos básicos del IPC y en las filas los períodos por meses. Las ponderaciones se obtienen de la estructura gb, en la tercera columna de ponderaciones.

Para conocer la estructura de los DataFrames necesarios, vea también: FullCPIBase.

CPIDataBase.InflationCombination — Type

InflationCombination <: EnsembleInflationFunction <: InflationFunction

Alias para CombinationFunction.

CPIDataBase.InflationEnsemble — Type

InflationEnsemble <: EnsembleInflationFunction <: InflationFunction

Alias para EnsembleFunction.

CPIDataBase.InflationFunction — Type

abstract type InflationFunction <: Function

Tipo abstracto para representar las funciones de inflación que operan sobre CountryStructure y VarCPIBase. Permiten computar la medida de ritmo inflacionario interanual, el índice de precios dado por la metodología y las variaciones intermensuales del índice de precios.

CPIDataBase.Item — Type

Item{T}(code::String, name::String, weight::T<:AbstractFloat)

Representa un gasto básico en la estructura de nodos del IPC. Es el nivel más bajo de la estructura. Almacena el código del gasto básico, su nombre o descripción y su ponderación en el IPC. Los datos de este nodo deben estar disponibles en algún FullCPIBase.

Posee los campos:

code: que almacena el código del gasto básico como un String.

name: que almacena el nombre del gasto básico como un String.

weight::T: que almacena la ponderación del gasto básico como un valor flotante de tipo T.

Aunque este nodo será usualmente creado de manera automática por métodos como CPITree, se pueden crear estructuras jerárquicas manualmente. Por ejemplo, para crear un nodo del nivel inferior:

julia> Item("_011101", "Item A", 7.352945f0)
Item{Float32}("_011101", "Item A", 7.352945f0)

Ver también: Group, CPITree.

CPIDataBase.MixedCountryStructure — Type

MixedCountryStructure{N, T} <: CountryStructure{N, T}

Estructura que representa el conjunto de bases del IPC de un país, posee el campo base, que es una tupla de la estructura VarCPIBase, cada una con su propio tipo de índices base B. Este tipo es una colección de un tipo abstracto.

CPIDataBase.UniformCountryStructure — Type

UniformCountryStructure{N, T, B} <: CountryStructure{N, T}

Estructura que representa el conjunto de bases del IPC de un país, posee el campo base, que es una tupla de la estructura VarCPIBase. Todas las bases deben tener el mismo tipo de índice base.

CPIDataBase.VarCPIBase — Type

VarCPIBase{T, B} <: AbstractCPIBase{T}

VarCPIBase(v::Matrix{T}, w::Vector{T}, dates::DATETYPE, baseindex::B) where {T, B}

Contenedor genérico para de variaciones intermensuales de índices de precios del IPC de un país. Se representa por:

Matriz de variaciones intermensuales v. En las filas contiene los períodos y en las columnas contiene los gastos básicos.

Vector de ponderaciones w de los gastos básicos.

Fechas correspondientes dates (por meses).

Índices base baseindex.

Este tipo es el utilizado en el contenedor de bases del IPC de un país, denominado CountryStructure, ya que con los datos de un VarCPIBase es suficiente para computar cualquier medida de inflación basada en índices de precios individuales o en un estadístico de resumen de las variaciones intermensuales (como un percentil, o una media truncada).

El tipo T representa el tipo de datos para representar los valores de punto flotante. El tipo B representa el tipo del campo baseindex; por ejemplo, Float32 o Vector{Float32}.

Ver también: CountryStructure, UniformCountryStructure, MixedCountryStructure

CPIDataBase.VarCPIBase — Method

VarCPIBase(df::DataFrame, gb::DataFrame)

Este constructor devuelve una estructura VarCPIBase a partir del DataFrame de índices de precios df, que contiene en las columnas las categorías o gastos básicos del IPC y en las filas los períodos por meses. Las ponderaciones se obtienen de la estructura gb, en la tercera columna de ponderaciones.

Para conocer la estructura de los DataFrames necesarios, vea también: FullCPIBase.

Base.eltype — Method

eltype(::CountryStructure{N, T})

Tipo de dato de punto flotante del contenedor de la estructura de país CountryStructure.

Base.getindex — Method

Base.getindex(cpitree::CPITree, code::AbstractString)

Este método se utiliza para indexar el árbol jerárquico cpitree y obtener una estructura similar cuyo nodo superior sea el nodo con el código provisto code. Por ejemplo, si tenemos el siguiente árbol:

julia> tree
CPITree{Group{Group{Group{Group{Item{Float32}, Float32}, Float32}, Float32}, Float32}} con datos
└─→ FullCPIBase{Float32, Float32}: 36 períodos × 10 gastos básicos Jan-01-Dec-03
_0: IPC [100.0]
├─ _01: Div._01 [21.491905]
│  └─ _011: Agr._011 [21.491905]
│     ├─ _0111: Subgr._0111 [7.352945]
│     │  └─ _011101: Item A [7.352945]
│     └─ _0112: Subgr._0112 [14.13896]
│        ├─ _011201: Item B [6.7442417]
│        └─ _011202: Item C [7.394718]
├─ _02: Div._02 [3.0530455]
│  ├─ _021: Agr._021 [1.1036392] 
│  │  └─ _0211: Subgr._0211 [1.1036392]
│  │     └─ _021101: Item D [1.1036392]
│  └─ _022: Agr._022 [1.9494063]
│     └─ _0221: Subgr._0221 [1.9494063]
│        └─ _022101: Item E [1.9494063]
├─ _03: Div._03 [11.68543]
│  └─ _031: Agr._031 [11.68543]
│     └─ _0311: Subgr._0311 [11.68543]
│        └─ _031101: Item F [11.68543]
└─ _04: Div._04 [63.769615]
   └─ _041: Agr._041 [63.769615]
      ├─ _0411: Subgr._0411 [16.103952]
      │  └─ _041101: Item G [16.103952]
      ├─ _0412: Subgr._0412 [28.824577]
      │  ├─ _041201: Item H [11.367162]
      │  └─ _041202: Item I [17.457417]
      └─ _0413: Subgr._0413 [18.841085]
         └─ _041301: Item J [18.841085]

Al indexar por un código, como _041, obtenemos una estructura similar a partir de ese nodo:

julia> tree["_041"]
CPITree{Group{Group{Item{Float32}, Float32}, Float32}} con datos
└─→ FullCPIBase{Float32, Float32}: 36 períodos × 10 gastos básicos Jan-01-Dec-03
_041: Agr._041 [63.769615]
├─ _0411: Subgr._0411 [16.103952]
│  └─ _041101: Item G [16.103952] 
├─ _0412: Subgr._0412 [28.824577]
│  ├─ _041201: Item H [11.367162]
│  └─ _041202: Item I [17.457417]
└─ _0413: Subgr._0413 [18.841085]
   └─ _041301: Item J [18.841085]

Base.getindex — Method

getindex(cst::CountryStructure, startdate::Date, finaldate::Date)

Devuelve una copia del CountryStructure con las bases modificadas para tener observaciones entre las fechas indicada por startdate y finaldate.

Base.getindex — Method

getindex(cst::CountryStructure, finaldate::Date)

Devuelve una copia del CountryStructure hasta la fecha indicada por finaldate.

Base.getindex — Method

getindex(cst::CountryStructure, i::Int)

Devuelve la base número i de un contenedor CountryStructure.

CPIDataBase.capitalize — Function

capitalize(v::AbstractVector, base_index::Real = 100)
capitalize(vmat::AbstractMatrix, base_index::Real = 100)
capitalize(vmat::AbstractMatrix, base_index::AbstractVector)

Función para encadenar un vector o matriz con variaciones intermensuales de índices de precios v o vmat para conformar un índice de precios cuyo valor base sea base_index.

CPIDataBase.capitalize! — Function

capitalize!(idx:: AbstractVector, v::AbstractVector, base_index::Real)
capitalize!(vmat::AbstractMatrix, base_index = 100)

Función para encadenar un vector o matriz con variaciones intermensuales de índices de precios v o vmat para conformar un índice de precios cuyo valor base sea base_index y sea almacenado en idx o en la propia matriz vmat.

CPIDataBase.capitalize — Method

capitalize(base::VarCPIBase)

Esto devuelve una nueva instancia (copia) de tipo IndexCPIBase de un objeto VarCPIBase.

CPIDataBase.components — Method

components(inflfn::CombinationFunction)

Devuelve un DataFrame con las componentes de la función de combinación lineal y las ponderaciones asociadas.

CPIDataBase.components — Method

components(inflfn::EnsembleFunction)

Devuelve un DataFrame con las componentes del EnsembleFunction.

CPIDataBase.compute_index — Method

compute_index(cpitree::CPITree [, code::AbstractString])

Permite computar el índice de precios de la jerarquía provista en code. Si se omite code, se computa la jerarquía padre de la estructura cpitree. Si el nodo no se encuentra en el árbol, devuelve nothing.

julia> tree
CPITree{Group{Group{Group{Group{Item{Float32}, Float32}, Float32}, Float32}, Float32}} con datos
└─→ FullCPIBase{Float32, Float32}: 36 períodos × 10 gastos básicos Jan-01-Dec-03
_0: IPC [100.0]
├─ _01: Div._01 [21.491905]
│  └─ _011: Agr._011 [21.491905]
│     ├─ _0111: Subgr._0111 [7.352945]
│     │  └─ _011101: Item A [7.352945]
│     └─ _0112: Subgr._0112 [14.13896]
│        ├─ _011201: Item B [6.7442417]
│        └─ _011202: Item C [7.394718]
├─ _02: Div._02 [3.0530455]
│  ├─ _021: Agr._021 [1.1036392]
│  │  └─ _0211: Subgr._0211 [1.1036392]
│  │     └─ _021101: Item D [1.1036392]
│  └─ _022: Agr._022 [1.9494063]
│     └─ _0221: Subgr._0221 [1.9494063]
│        └─ _022101: Item E [1.9494063]
├─ _03: Div._03 [11.68543]
│  └─ _031: Agr._031 [11.68543]
│     └─ _0311: Subgr._0311 [11.68543]
│        └─ _031101: Item F [11.68543]
└─ _04: Div._04 [63.769615]
   └─ _041: Agr._041 [63.769615]
      ├─ _0411: Subgr._0411 [16.103952]
      │  └─ _041101: Item G [16.103952]
      ├─ _0412: Subgr._0412 [28.824577]
      │  ├─ _041201: Item H [11.367162]
      │  └─ _041202: Item I [17.457417]
      └─ _0413: Subgr._0413 [18.841085]
         └─ _041301: Item J [18.841085]

julia> compute_index(tree, "_041")
36-element Vector{Float32}:
 100.13899
 100.2787
 100.41912
 100.560234
 100.70207
 100.84464
 100.98792
   ⋮
 104.65377
 104.81638
 104.97984
 105.14412
 105.309235
 105.47519

julia> compute_index(tree["_041"])
36-element Vector{Float32}:
 100.13899
 100.2787
 100.41912
 100.560234
 100.70207
 100.84464
 100.98792
   ⋮
 104.65377
 104.81638
 104.97984
 105.14412
 105.309235
 105.47519

julia> a = tree["_041302"]

julia> a === nothing
true

CPIDataBase.cpi_tree_nodes — Method

cpi_tree_nodes(codes::Vector{<:AbstractString}; 
    characters::(NTuple{N, Int} where N), depth::Int=1, chars::Int=characters[depth], prefix::AbstractString="", 
    base::FullCPIBase, 
    group_names::Vector{<:AbstractString}, 
    group_codes::Vector{<:AbstractString})

Construye y devuelve una lista de nodos a partir de la lista de códigos codes o de la especificación jerárquica de caracteres en characters. Los nombres y las ponderaciones del nivel inferior (nivel de gasto básico) son obtenidas de la estructura base. Se debe proveer el vector de códigos y nombres de todas las jerarquías superiores en la estructura de códigos en los vectores group_names y group_codes.

Esta función permite crear únicamente la estructura jerárquica de nodos. Para construir de manera automática una estructura del IPC, se recomienda utilizar preferentemente CPITree.

Vea también: CPITree.

CPIDataBase.get_cpi_tree — Method

get_cpi_tree(; 
    base::FullCPIBase, 
    group_names::Vector{<:AbstractString}, 
    group_codes::Vector{<:AbstractString}, 
    characters::(NTuple{N, Int} where N),
    upperlevel_code = "_0", 
    upperlevel_name = "IPC")

Función superior para obtener estructura jerárquica del IPC. Devuelve el nodo superior del árbol jerárquico. Utiliza la función de más bajo nivel cpi_tree_nodes para construir los nodos del nivel más alto y hacia abajo en la estructura jerárquica. Se debe proveer el vector de códigos y nombres de todas las jerarquías superiores en la estructura de códigos en los vectores group_names y group_codes.

Esta función permite crear únicamente la estructura jerárquica de nodos. Para construir de manera automática una estructura del IPC, se recomienda utilizar preferentemente CPITree.

Vea también: CPITree.

CPIDataBase.getdates — Method

getdates(startdate::Date, vmat::AbstractMatrix)

Obtiene un rango de fechas a partir de una fecha inicial startdate y la cantidad de períodos en la matriz de variaciones intermensuales vmat.

CPIDataBase.getdates — Method

getdates(startdate::Date, periods::Int)

Obtiene un rango de fechas a partir de una fecha inicial startdate y la cantidad de períodos de una matriz de variaciones intermensuales

CPIDataBase.getunionalltype — Method

getunionalltype(::MixedCountryStructure)

Devuelve el tipo MixedCountryStructure. Utilizado al llamar getunionalltype sobre un CountryStructure para obtener el tipo concreto UnionAll.

CPIDataBase.getunionalltype — Method

getunionalltype(::UniformCountryStructure)

Devuelve el tipo UniformCountryStructure. Utilizado al llamar getunionalltype sobre un CountryStructure para obtener el tipo concreto UnionAll.

CPIDataBase.infl_dates — Method

infl_periods(cst::CountryStructure)

Fechas correspondientes a la trayectorias de inflación computadas a partir un CountryStructure.

CPIDataBase.infl_periods — Method

infl_periods(cst::CountryStructure)

Computa el número de períodos de inflación de la estructura de país. Corresponde al número de observaciones intermensuales menos las primeras 11 observaciones de la primera base del IPC.

CPIDataBase.measure_name — Method

measure_name(inflfn::InflationFunction)

Este método permite obtener el nombre convencional de una medida de inflación.

CPIDataBase.measure_tag — Method

measure_tag(inflfn::InflationFunction)

Obtiene una etiqueta de la medida de inflación. Se puede utilizar para guardar como parámetro en archivos de resultados de evaluación.

CPIDataBase.num_measures — Method

num_measures(::InflationFunction)

Devuelve la cantidad de medidas devueltas por la función de inflación. Las funciones de EnsembleFunction pueden computar varias medidas de inflación simultáneamente.

CPIDataBase.params — Method

params(inflfn::InflationFunction)

Método para obtener parámetros de la función de inflación. Devuelve una tupla con el conjunto de parámetros utilizado por la función de inflación inflfn. Este método debe redefinirse en las nuevas medidas de inflación si estas están parametrizadas.

CPIDataBase.periods — Method

periods(cst::CountryStructure)

Computa el número de períodos (meses) en las bases de variaciones intermensuales de la estructura de país.

CPIDataBase.periods — Method

periods(base::VarCPIBase)

Computa el número de períodos (meses) en las base de variaciones intermensuales.

CPIDataBase.varinteran — Function

varinteran(idx::AbstractVector, base_index::Real = 100) -> Vector{<:AbstractFloat}
varinteran(cpimat::AbstractMatrix, base_index::Real = 100) -> Matrix{<:AbstractFloat}
varinteran(cpimat::AbstractMatrix, base_index::AbstractVector) -> Matrix{<:AbstractFloat}

Obtiene variaciones interanuales del vector idx o de la matriz cpimat utilizando como índice base el número o vector base_index.

Si base_index es un vector, se obtienen las variaciones interanuales utilizando diferentes índices base para cada columna de cpimat. El vector base_index debe tener la misma cantidad de columnas que cpimat.

CPIDataBase.varinteran! — Function

varinteran!(v::AbstractVector, idx::AbstractVector, base_index::Real = 100) -> Vector{<:AbstractFloat}

Computa las variaciones interanuales del vector idx utilizando como índice base base_index. Si se provee v, los resultados son guardados en este vector, en vez del mismo idx.

El vector v tiene 11 observaciones menos que idx.

CPIDataBase.varinterm — Function

varinterm(idx::AbstractVecOrMat, base_index = 100)

Función para computar un vector o una matriz de variaciones intermensuales de los índices de precios en idx, utilizando como índice base base_index en la primera observación.

Ver también: varinterm!

CPIDataBase.varinterm! — Function

varinterm!([v::AbstractVecOrMat, ] idx::AbstractVecOrMat, base_index = 100)

Función para computar un vector o una matriz de variaciones intermensuales de los índices de precios en idx, utilizando como índice base base_index en la primera observación. Si idx es una matriz, v es opcional y el cómputo se realiza sobre la misma matriz. Si idx es un vector, es necesario proporcionar v para realizar el cómputo.

Ver también: varinterm.

CPIDataBase.varinterm — Method

varinterm(base::IndexCPIBase)

Devuelve una nueva copia de tipo VarCPIBase de un IndexCPIBase.

CPIDataBase.weights — Method

weights(combfn::CombinationFunction)

Devuelve el vector de ponderaciones de una CombinationFunction.

CPIDataBase.TestHelpers.getbasedates — Function

getbasedates(vmat, startdate=Date(2000, 12))

Función para generar fechas a partir de matriz de variaciones intermensuales

CPIDataBase.TestHelpers.getrandombase — Method

getrandombase(; 
    [T_type = Float32, 
    G = 218, 
    T_periods = 120, 
    startdate = Date(2001,1), 
    baseindex = 100*one(T_type)])

Función para obtener una base de tipo VarCPIBase con variaciones intermensuales aleatorias.

CPIDataBase.TestHelpers.getrandomcountryst — Function

getrandomcountryst(T_type = Float32)

Obtiene un UniformCountryStructure cuyas variaciones intermensuales son todas aleatorias en la configuración de períodos del IPC de Guatemala.

CPIDataBase.TestHelpers.getrandomweights — Function

getrandomweights(T=Float32, G=218)

Función para generar pesos aleatorios

CPIDataBase.TestHelpers.getzerobase — Method

getzerobase(; 
    [T_type = Float32, 
    G = 218, 
    T_periods = 120, 
    startdate = Date(2001,1), 
    baseindex = 100*one(T_type)])

Función para obtener base de tipo VarCPIBase con variaciones intermensuales iguales a cero.

CPIDataBase.TestHelpers.getzerocountryst — Function

getzerocountryst(T_type = Float32)

Obtiene un UniformCountryStructure cuyas variaciones intermensuales son todas iguales a cero en la configuración de períodos del IPC de Guatemala.