HEMI

HEMI.HEMI — Module

HEMI

Módulo envolvente que carga los paquetes y datos utilizados en todo el proyecto.

HEMI.load_data — Method

load_data(; full_precision = false)

Carga los datos del archivo principal de datos HEMI.datafile. Si los datos no están presentes, se cargan datos de prueba del archivo incluido por defecto, HEMI.testdatafile.

La opción full_precision permite cargar datos con precisión de 64 bits.

Archivo principal: HEMI.datafile = datadir("guatemala", "gtdata32.jld2").

Archivo principal (64 bits): HEMI.datafile = datadir("guatemala", "gtdata.jld2").

Utilización en scripts de evaluación

Este módulo se utiliza para cargar los paquetes principales utilizados en la evaluación de medidas de inflación.

Si ya se ha ejecutado el script de carga de datos scripts/load_data.jl, este módulo carga los datos del IPC de Guatemala en el objeto gtdata, el cual es un objeto de tipo UniformCountryStructure. Provee además los objetos gt00 y gt10, los cuales son de tipo VarCPIBase.

Como se describe en la documentación de DrWatson, existen proyectos en los que se deben cargar datos y funciones al inicio de cualquier archivo del proyecto. La estructura del proyecto en Julia permite englobar los paquetes más importantes en un módulo con el mismo nombre del proyecto. Por lo tanto, podemos utilizar el siguiente código al inicio de cada script para activar el proyecto y cargar los paquetes principales:

using DrWatson 
@quickactivate :HEMI

## Script de evaluación o pruebas
# ...

el cual es equivalente a

using DrWatson 
@quickactivate "HEMI"
using HEMI

## Script de evaluación o pruebas
# ...

Ejemplo para computar trayectoria de inflación

Ejecutar el archivo scripts/load_data.jl para generar el archivo binario de datos en formato JLD2.

Crear un archivo de prueba en scripts llamado testHEMI.jl e incluir el siguiente código para computar la variación interanual del IPC.

## Activar el entorno del proyecto y cargar los datos
using DrWatson
@quickactivate :HEMI

## Utilizar la función de inflación InflationTotalCPI sobre gtdata
totalfn = InflationTotalCPI()
traj_infl = totalfn(gtdata)