InflationEvalTools

InflationEvalTools.InflationEvalTools

InflationEvalTools.LOCAL_RNG

InflationEvalTools.RWTREND

InflationEvalTools.AbstractInflationParameter

InflationEvalTools.ArrayTrendFunction

InflationEvalTools.InflationParameter

InflationEvalTools.ResampleFunction

InflationEvalTools.ResampleFunction

InflationEvalTools.ResampleFunction

InflationEvalTools.ResampleScrambleVarMonths

InflationEvalTools.ResampleScrambleVarMonths

InflationEvalTools.ScrambleVarMonths

InflationEvalTools.TrendAnalytical

InflationEvalTools.TrendFunction

InflationEvalTools.TrendIdentity

InflationEvalTools.TrendRandomWalk

InflationEvalTools.DEFAULT_RESAMPLE_FN

InflationEvalTools.ParamTotalCPI

InflationEvalTools.ParamTotalCPIRebase

InflationEvalTools.ParamWeightedMean

InflationEvalTools.gentrayinfl

InflationEvalTools.get_param_function

InflationEvalTools.get_ranges

InflationEvalTools.method_name

InflationEvalTools.method_tag

InflationEvalTools.param_gsbb_mod

InflationEvalTools.param_gsbb_mod

InflationEvalTools.param_sbb

InflationEvalTools.param_sbb

InflationEvalTools.pargentrayinfl

InflationEvalTools.scramblevar

InflationEvalTools.scramblevar!

InflationEvalTools.InflationEvalTools — Module

InflationEvalTools

Tipos, funciones y demás utilidades de simulación para evaluación de medidas inflación.

InflationEvalTools.LOCAL_RNG — Constant

Esta constante se utiliza para fijar el generador de números aleatorios en cada proceso al generador MersenneTwister con semilla inicial cero.

InflationEvalTools.RWTREND — Constant

RWTREND

Trayectoria de caminata aleatoria precalibrada para 240 períodos.

InflationEvalTools.AbstractInflationParameter — Type

Tipo abstracto para representar los parámetros de inflación

InflationEvalTools.ArrayTrendFunction — Type

abstract type ArrayTrendFunction <: TrendFunction end

Tipo para función de tendencia que almacena el vector de valores a aplicar a las variaciones intermensuales.

Utilización

(trendfn::ArrayTrendFunction)(base::VarCPIBase{T}, range::UnitRange) where T

Especifica cómo aplicar la función de tendencia sobre un VarCPIBase con el rango de índices range.

InflationEvalTools.InflationParameter — Type

Tipo concreto para representar un parámetro de inflación computado con la función de inflación inflfn y el método de remuestreo resamplefn.

Ver también: ParamTotalCPIRebase, ParamTotalCPI, ParamWeightedMean

InflationEvalTools.ResampleFunction — Type

abstract type ResampleFunction <: Function end

Tipo abstracto para funciones de remuestreo. Cada función debe extender como mínimo el método

resamplefn(vmat::AbstractMatrix, rng)::Matrix

para remuestrear un CountryStructure con las funciones definidas arriba.

Opcionalmente, si se desea modificar el comportamiento específico de cada función de remuestreo, se deben extender los siguientes métodos:

function (resamplefn::ResampleFunction)(cs::CountryStructure, rng = Random.GLOBAL_RNG)

function (resamplefn::ResampleFunction)(base::VarCPIBase, rng = Random.GLOBAL_RNG)

InflationEvalTools.ResampleFunction — Type

function (resamplefn::ResampleFunction)(cs::CountryStructure, rng = Random.GLOBAL_RNG)

Define el comportamiento general de función de remuestreo sobre CountryStructure. Se remuestrea cada una de las bases del campo base utilizando el método para objetos VarCPIBase y se devuelve un nuevo CountryStructure.

InflationEvalTools.ResampleFunction — Type

function (resamplefn::ResampleFunction)(base::VarCPIBase, rng = Random.GLOBAL_RNG)

Define el comportamiento general de función de remuestreo sobre VarCPIBase. Este método requiere una implementación específica del método sobre el par (AbstractMatrix, rng). Considera que el método de remuestreo podría extender los períodos de la serie de tiempo y ajusta las fechas apropiadamente.

InflationEvalTools.ResampleScrambleVarMonths — Type

ResampleScrambleVarMonths

Alias para InflationEvalTools.ScrambleVarMonths

InflationEvalTools.ResampleScrambleVarMonths — Type

scramblevar(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG)

Define cómo remuestrear matrices con las series de tiempo en las columnas. Utiliza la función interna scramblevar.

InflationEvalTools.ScrambleVarMonths — Type

struct ScrambleVarMonths <: ResampleFunction end

Define una función de remuestreo para remuestrear las series de tiempo por los mismos meses de ocurrencia. El muestreo se realiza de manera independiente para serie de tiempo en las columnas de una matriz.

InflationEvalTools.TrendAnalytical — Type

TrendAnalytical{T} <: ArrayTrendFunction

Tipo para representar una función de tendencia definida por una función anónima. Recibe los datos de un CountryStructure o un rango de índices para precomputar el vector de tendencia utlizando una función anónima.

Ejemplos:

Para crear una función de tendencia a partir de una función anónima:

trendfn = TrendAnalytical(param_data, t -> 1 + sin(2π*t/12))

o bien:

trendfn = TrendAnalytical(1:periods(param_data), t -> 1 + sin(2π*t/12))

InflationEvalTools.TrendFunction — Type

abstract type TrendFunction <: Function end

Tipo abstracto para manejar las funciones de tendencia.

Utilización

(trendfn::TrendFunction)(cs::CountryStructure)

Aplica la función de tendencia sobre un CountryStructure y devuelve un nuevo CountryStructure.

InflationEvalTools.TrendIdentity — Type

TrendIdentity <: TrendFunction

Tipo concreto para representar una función de tendencia neutra. Es decir, esta función de tendencia mantiene los datos sin alteración.

Ejemplos:

# Crear una función de tendencia neutra. 
trendfn = TrendIdentity()

Utilización

(trendfn::TrendIdentity)(cs::CountryStructure)

Aplicación de tendencia TrendIdentity sobre VarCPIBase. Se redefine este método

para dejar invariante la base VarCPIBase.

trendfn = TrendIdentity() 
trended_cs = trendfn(gtdata) 

InflationEvalTools.TrendRandomWalk — Type

TrendRandomWalk{T} <: ArrayTrendFunction

Tipo para representar una función de tendencia de caminata aleatoria. Utiliza el vector de caminata aleatoria precalibrado en RWTREND.

Ejemplo:

# Crear la función de tendencia de caminata aleatoria
trendfn = TrendRandomWalk()

InflationEvalTools.DEFAULT_RESAMPLE_FN — Function

DEFAULT_RESAMPLE_FN

Define la funcón de remuestreo a utilizar por defecto en el ejercicio de simulación.

InflationEvalTools.ParamTotalCPI — Method

ParamTotalCPI()

Función de ayuda para obtener la configuración del parámetro de inflación dado por la función de inflación del IPC y el método de remuestreo por defecto.

InflationEvalTools.ParamTotalCPIRebase — Method

ParamTotalCPIRebase()

Función de ayuda para obtener la configuración del parámetro de inflación dado por la función de inflación del IPC con cambio de base sintético y el método de remuestreo por defecto.

InflationEvalTools.ParamWeightedMean — Method

ParamWeightedMean()

Función de ayuda para obtener la configuración del parámetro de inflación dado por la media ponderada interanual y el método de remuestreo por defecto.

InflationEvalTools.gentrayinfl — Method

gentrayinfl(inflfn::F, resamplefn::R, trendfn::T, csdata::CountryStructure; K = 100, rndseed = 0, showprogress = true)

Computa K trayectorias de inflación utilizando la función de inflación inflfn::InflationFunction, la función de remuestreo resamplefn::TrendFunction y la función de tendencia trendfn::TrendFunction especificada. Se utilizan los datos en el CountryStructure dado en csdata.

A diferencia de la función pargentrayinfl, esta función no realiza el cómputo de forma distribuida.

Para lograr la reproducibilidad entre diferentes corridas de la función, y de esta forma, generar trayectorias de inflación con diferentes metodologías utilizando los mismos remuestreos, se fija la semilla de generación de acuerdo con el número de iteración en la simulación. Para controlar el inicio de la generación de trayectorias se utiliza como parámetro de desplazamiento el valor rndseed.

InflationEvalTools.get_param_function — Method

get_param_function(::ResampleFunction)

Función para obtener de una función de remuestreo la función que permite obtener las variaciones intermensuales promedio (o paramétricas), que finalmente sirven para construir la trayectoria paramétrica de inflación de la metodología de remuestreo.

InflationEvalTools.get_ranges — Method

get_ranges(cs::CountryStructure)

Función de ayuda para obtener tupla de rangos de índices para hacer slicing de los vectores de tendencia.

InflationEvalTools.method_name — Method

method_name(resamplefn::ResampleFunction)

Función para obtener el nombre del método de remuestreo.

InflationEvalTools.method_tag — Method

method_tag(resamplefn::ResampleFunction)

Función para obtener una etiqueta del método de remuestreo.

InflationEvalTools.param_gsbb_mod — Method

param_gsbb_mod(cs::CountryStructure)

Obtiene un CountryStructure paramétrico.

InflationEvalTools.param_gsbb_mod — Method

param_gsbb_mod(base::VarCPIBase)

Obtiene la matriz de variaciones intermensuales paramétricas para la metodología de remuestreo de Generalized Seasonal Block Bootstrap modificada que extiende las observaciones a 300 períodos. Devuelve una base de tipo VarCPIBase con las variaciones intermensuales paramétricas. Actualmente funciona solamente si base tiene 120 observaciones.

InflationEvalTools.param_sbb — Method

param_sbb(cs::CountryStructure)

Obtiene un CountryStructure paramétrico. Véase también param_sbb.

InflationEvalTools.param_sbb — Method

param_sbb(base::VarCPIBase)

Obtiene la matriz de variaciones intermensuales paramétricas para la metodología de remuestreo de Stationary Block Bootstrap. Devuelve una base de tipo VarCPIBase con las variaciones intermensuales promedio de los mismos meses de ocurrencia (también llamadas variaciones intermensuales paramétricas).

Esta definición también aplica a otras metodologías que utilicen como variaciones intermensuales paramétricas los promedios en los mismos meses de ocurrencia.

InflationEvalTools.pargentrayinfl — Method

pargentrayinfl(inflfn::F, resamplefn::R, trendfn::T, csdata::CountryStructure; K = 100, rndseed = 0, showprogress = true)

Computa K trayectorias de inflación utilizando la función de inflación inflfn::InflationFunction, la función de remuestreo resamplefn::TrendFunction y la función de tendencia trendfn::TrendFunction especificada. Se utilizan los datos en el CountryStructure dado en csdata.

A diferencia de la función gentrayinfl, esta función implementa el cómputo distribuido en procesos utilizando @distributed. Esto requiere que el paquete haya sido cargado en todos los procesos de cómputo. Por ejemplo:

using Distributed
addprocs(4, exeflags="--project")
@everywhere begin
    using HEMI 
    using InflationEvalTools
end

Para lograr la reproducibilidad entre diferentes corridas de la función, y de esta forma, generar trayectorias de inflación con diferentes metodologías utilizando los mismos remuestreos, se fija la semilla de generación de acuerdo con el número de iteración en la simulación. Para controlar el inicio de la generación de trayectorias se utiliza como parámetro de desplazamiento el valor rndseed.

InflationEvalTools.scramblevar — Function

scramblevar(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG)

Copy and scramble every column of matrix vmat by months.

InflationEvalTools.scramblevar! — Function

scramblevar!(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG)

Samples in-place every column of matrix vmat by months.