CPIDataBase

CPIDataBase.CPIDataBase

CPIDataBase.CPIIndex

CPIDataBase.CPIResult

CPIDataBase.CPIVarInterm

CPIDataBase.CombinationFunction

CPIDataBase.CountryStructure

CPIDataBase.EnsembleFunction

CPIDataBase.EnsembleInflationFunction

CPIDataBase.FullCPIBase

CPIDataBase.FullCPIBase

CPIDataBase.IndexCPIBase

CPIDataBase.IndexCPIBase

CPIDataBase.InflationFunction

CPIDataBase.MixedCountryStructure

CPIDataBase.UniformCountryStructure

CPIDataBase.VarCPIBase

CPIDataBase.VarCPIBase

CPIDataBase.capitalize

CPIDataBase.capitalize

CPIDataBase.capitalize

CPIDataBase.capitalize

CPIDataBase.capitalize!

CPIDataBase.capitalize!

CPIDataBase.capitalize_addbase

CPIDataBase.getdates

CPIDataBase.getdates

CPIDataBase.getunionalltype

CPIDataBase.getunionalltype

CPIDataBase.infl_dates

CPIDataBase.infl_periods

CPIDataBase.measure_name

CPIDataBase.measure_tag

CPIDataBase.num_measures

CPIDataBase.periods

CPIDataBase.periods

CPIDataBase.varinteran

CPIDataBase.varinteran

CPIDataBase.varinteran

CPIDataBase.varinteran!

CPIDataBase.varinteran!

CPIDataBase.varinterm

CPIDataBase.varinterm

CPIDataBase.varinterm

CPIDataBase.varinterm

CPIDataBase.varinterm!

CPIDataBase.varinterm!

CPIDataBase.weights

CPIDataBase.TestHelpers.getbasedates

CPIDataBase.TestHelpers.getrandomweights

CPIDataBase.TestHelpers.getzerobase

CPIDataBase.TestHelpers.getzerocountryst

CPIDataBase.CPIDataBase — Module

CPIDataBase

Librería base para tipos y funcionalidad básica para manejo de datos del IPC a nivel desagregado de gastos básicos

CPIDataBase.CPIIndex — Type

Tipo concreto único para obtener el índice de una función de inflación

CPIDataBase.CPIResult — Type

Tipo abstracto para manejar el despacho de las funciones de inflación

CPIDataBase.CPIVarInterm — Type

Tipo concreto único para obtener la variación intermensual de una función de inflación

CPIDataBase.CombinationFunction — Type

CombinationFunction{N, W} <: EnsembleInflationFunction

Función de inflación para computar un promedio ponderado de un conjunto de N de medidas de inflación con tipo del vector de ponderaciones W.

CPIDataBase.CountryStructure — Type

CountryStructure{N, T <: AbstractFloat}

Tipo abstracto que representa el conjunto de bases del IPC de un país.

CPIDataBase.EnsembleFunction — Type

EnsembleFunction{N} <: EnsembleInflationFunction

Función de inflación para computar un conjunto de N de medidas de inflación.

CPIDataBase.EnsembleInflationFunction — Type

EnsembleInflationFunction <: InflationFunction

Tipo abstracto para conformar conjuntos de medidas de inflación

CPIDataBase.FullCPIBase — Type

FullCPIBase{T, B} <: AbstractCPIBase{T}

Contenedor completo para datos del IPC de un país. Se representa por:

Matriz de índices de precios ipc que incluye la fila con los índices del númbero base.

Matriz de variaciones intermensuales v. En las filas contiene los períodos y en las columnas contiene los gastos básicos.

Vector de ponderaciones w de los gastos básicos.

Fechas correspondientes fechas (por meses).

CPIDataBase.FullCPIBase — Method

FullCPIBase(df::DataFrame, gb::DataFrame)

Este constructor devuelve una estructura FullCPIBase a partir del DataFrame de índices de precios df, que contiene en las columnas las categorías o gastos básicos del IPC y en las filas los períodos por meses. Las ponderaciones se obtienen de la estructura gb, en la columna denominada :Ponderacion.

CPIDataBase.IndexCPIBase — Type

IndexCPIBase{T, B} <: AbstractCPIBase{T}

Contenedor genérico de índices de precios del IPC de un país. Se representa por:

Matriz de índices de precios ipc que incluye la fila con los índices del númbero base.

Vector de ponderaciones w de los gastos básicos.

Fechas correspondientes fechas (por meses).

CPIDataBase.IndexCPIBase — Method

IndexCPIBase(df::DataFrame, gb::DataFrame)

Este constructor devuelve una estructura IndexCPIBase a partir del DataFrame de índices de precios df, que contiene en las columnas las categorías o gastos básicos del IPC y en las filas los períodos por meses. Las ponderaciones se obtienen de la estructura gb, en la columna denominada :Ponderacion.

CPIDataBase.InflationFunction — Type

abstract type InflationFunction <: Function

Tipo abstracto para representar las funciones de inflación que operan sobre CountryStructure y VarCPIBase. Permiten computar la medida de ritmo inflacionario interanual, el índice de precios dado por la metodología y las variaciones intermensuales del índice de precios.

CPIDataBase.MixedCountryStructure — Type

MixedCountryStructure{N, T} <: CountryStructure{N, T}

Estructura que representa el conjunto de bases del IPC de un país, posee el campo base, que es una tupla de la estructura VarCPIBase, cada una con su propio tipo de índices base B. Este tipo es una colección de un tipo abstracto.

CPIDataBase.UniformCountryStructure — Type

UniformCountryStructure{N, T, B} <: CountryStructure{N, T}

Estructura que representa el conjunto de bases del IPC de un país, posee el campo base, que es una tupla de la estructura VarCPIBase. Todas las bases deben tener el mismo tipo de índice base.

CPIDataBase.VarCPIBase — Type

VarCPIBase{T, B} <: AbstractCPIBase{T}

Contenedor genérico para de variaciones intermensuales de índices de precios del IPC de un país. Se representa por:

Matriz de variaciones intermensuales v. En las filas contiene los períodos y en las columnas contiene los gastos básicos.

Vector de ponderaciones w de los gastos básicos.

Fechas correspondientes fechas (por meses).

CPIDataBase.VarCPIBase — Method

VarCPIBase(df::DataFrame, gb::DataFrame)

Este constructor devuelve una estructura VarCPIBase a partir del DataFrame de índices de precios df, que contiene en las columnas las categorías o gastos básicos del IPC y en las filas los períodos por meses. Las ponderaciones se obtienen de la estructura gb, en la columna denominada :Ponderacion.

Base.eltype — Method

eltype(::CountryStructure{N, T})

Tipo de dato de punto flotante del contenedor de la estructura de país CountryStructure.

Base.getindex — Method

getindex(cst::CountryStructure, startdate::Date, finaldate::Date)

Devuelve una copia del CountryStructure con las bases modificadas para tener observaciones entre las fechas indicada por startdate y finaldate.

Base.getindex — Method

getindex(cst::CountryStructure, finaldate::Date)

Devuelve una copia del CountryStructure hasta la fecha indicada por finaldate.

Base.getindex — Method

getindex(cst::CountryStructure, i::Int)

Devuelve la base número i de un contenedor CountryStructure.

CPIDataBase.capitalize — Function

capitalize(v::AbstractVector, base_index::Real = 100)

Function to chain a vector of price changes with an index starting with base_index.

CPIDataBase.capitalize — Function

capitalize(vmat::AbstractMatrix, base_index::Real = 100)

Function to chain a matrix of price changes with an index starting with base_index.

CPIDataBase.capitalize! — Function

capitalize!(vmat::AbstractMatrix, base_index = 100)

Function to chain a matrix of price changes in place with an index starting with base_index.

CPIDataBase.capitalize! — Method

capitalize!(idx:: AbstractVector, v::AbstractVector, base_index::Real)

Function to chain a vector of price changes in vector idx with an index starting with base_index.

CPIDataBase.capitalize — Method

capitalize(vmat::AbstractMatrix, base_index::AbstractVector)

Function to chain a matrix of price changes with an index vector starting with base_index. base_index needs to have the same number of elements as the columns of vmat.

CPIDataBase.capitalize — Method

capitalize(base::VarCPIBase)

This returns a new instance (copy) of type IndexCPIBase from a VarCPIBase.

CPIDataBase.capitalize_addbase — Function

capitalize_addbase(vmat::AbstractMatrix, base_index = 100)

Function to chain a matrix of price changes with an index starting with base_index. This function adds the base index as the first row of the returned matrix.

CPIDataBase.getdates — Method

getdates(startdate::Date, vmat::AbstractMatrix)

Obtiene un rango de fechas a partir de una fecha inicial startdate y la cantidad de períodos en la matriz de variaciones intermensuales vmat.

CPIDataBase.getdates — Method

getdates(startdate::Date, periods::Int)

Obtiene un rango de fechas a partir de una fecha inicial startdate y la cantidad de períodos de una matriz de variaciones intermensuales

CPIDataBase.getunionalltype — Method

getunionalltype(::MixedCountryStructure)

Devuelve el tipo MixedCountryStructure. Utilizado al llamar getunionalltype sobre un CountryStructure para obtener el tipo concreto UnionAll.

CPIDataBase.getunionalltype — Method

getunionalltype(::UniformCountryStructure)

Devuelve el tipo UniformCountryStructure. Utilizado al llamar getunionalltype sobre un CountryStructure para obtener el tipo concreto UnionAll.

CPIDataBase.infl_dates — Method

infl_periods(cst::CountryStructure)

Fechas correspondientes a la trayectorias de inflación computadas a partir un CountryStructure.

CPIDataBase.infl_periods — Method

infl_periods(cst::CountryStructure)

Computa el número de períodos de inflación de la estructura de país. Corresponde al número de observaciones intermensuales menos las primeras 11 observaciones de la primera base del IPC.

CPIDataBase.measure_name — Method

measure_name(inflfn::InflationFunction)

Este método permite obtener el nombre convencional de una medida de inflación.

CPIDataBase.measure_tag — Method

measure_tag(inflfn::InflationFunction)

Obtiene una etiqueta de la medida de inflación. Se puede utilizar para guardar como parámetro en archivos de resultados de evaluación.

CPIDataBase.num_measures — Method

num_measures(::InflationFunction)

Devuelve la cantidad de medidas devueltas por la función de inflación. Las funciones de EnsembleFunction pueden computar varias medidas de inflación simultáneamente.

CPIDataBase.periods — Method

periods(cst::CountryStructure)

Computa el número de períodos (meses) en las bases de variaciones intermensuales de la estructura de país.

CPIDataBase.periods — Method

periods(base::VarCPIBase)

Computa el número de períodos (meses) en las base de variaciones intermensuales.

CPIDataBase.varinteran — Function

varinteran(idx::AbstractVector, base_index::Real = 100)

Function to get a vector of annual price changes from a price index vector starting with base_index.

CPIDataBase.varinteran — Function

varinteran(cpimat::AbstractMatrix, base_index::Real = 100)

Function to get a matrix of annual price changes from a price index matrix starting with base_index.

CPIDataBase.varinteran! — Function

varinteran!(idx, base_index::Real = 100)

Compute annual price changes of idx index vector in place, using base_index as starting point. Fills observations 1 to 11 with NaN values.

CPIDataBase.varinteran! — Function

varinteran!(v, idx, base_index::Real = 100)

Fill v vector of annual price changes of idx index vector using base_index as starting point. Usually v has 11 observations less than idx.

CPIDataBase.varinteran — Method

varinteran(cpimat::AbstractMatrix, base_index::AbstractVector)

Function to get a matrix of annual price changes from a price index matrix starting with vector base_index.

CPIDataBase.varinterm — Function

varinterm(idx::AbstractVector, base_index::Real = 100)

Function to get a vector of price changes from a price index vector starting with base_index.

CPIDataBase.varinterm — Function

varinterm(cpimat::AbstractMatrix, base_index::Real = 100)

Function to get a matrix of price changes from a price index matrix starting with base_index.

CPIDataBase.varinterm! — Function

varinterm!(v, idx, base_index::Real = 100)

Fill v vector of price changes of idx vector using base_index as starting point.

CPIDataBase.varinterm! — Function

varinterm!(cpimat::AbstractMatrix, base_index::Real = 100)

Function to get a matrix of price changes from a price index matrix starting with base_index.

CPIDataBase.varinterm — Method

varinterm(cpimat::AbstractMatrix, base_index::AbstractVector)

Function to get a matrix of price changes from a price index matrix starting with vector base_index.

CPIDataBase.varinterm — Method

varinterm(base::IndexCPIBase)

This returns a new instance (copy) of type VarCPIBase from an IndexCPIBase.

CPIDataBase.weights — Method

weights(combfn::CombinationFunction)

Devuelve el vector de ponderaciones de una CombinationFunction.

CPIDataBase.TestHelpers.getbasedates — Function

getbasedates(vmat, startdate=Date(2000, 12))

Función para generar fechas a partir de matriz de variaciones intermensuales

CPIDataBase.TestHelpers.getrandomweights — Function

getrandomweights(T=Float32, G=218)

Función para generar pesos aleatorios

CPIDataBase.TestHelpers.getzerobase — Function

get_zero_base(T_type=Float32, G=218, T_periods=120, startdate=Date(2001,1), baseindex=100*one(T_type))

Función para obtener base de tipo VarCPIBase con variaciones intermensuales iguales a cero.

CPIDataBase.TestHelpers.getzerocountryst — Function

getzerocountryst(T_type=Float32)

Obtiene un UniformCountryStructure cuyas variaciones intermensuales son todas iguales a cero en la configuración de períodos del IPC de Guatemala.