InflationEvalTools

InflationEvalTools.InflationEvalTools

InflationEvalTools.ResampleFunction

InflationEvalTools.ResampleFunction

InflationEvalTools.ResampleFunction

InflationEvalTools.ResampleScrambleVarMonths

InflationEvalTools.ResampleScrambleVarMonths

InflationEvalTools.ScrambleVarMonths

InflationEvalTools.apply_trend

InflationEvalTools.get_param_function

InflationEvalTools.method_name

InflationEvalTools.param_gsbb_mod

InflationEvalTools.param_gsbb_mod

InflationEvalTools.param_sbb

InflationEvalTools.param_sbb

InflationEvalTools.scramblevar

InflationEvalTools.scramblevar!

InflationEvalTools.InflationEvalTools — Module

InflationEvalTools

Tipos, funciones y demás utilidades de simulación para evaluación de medidas inflación.

InflationEvalTools.ResampleFunction — Type

abstract type ResampleFunction <: Function end

Tipo abstracto para funciones de remuestreo. Cada función debe extender como mínimo el método

resamplefn(vmat::AbstractMatrix, rng)::Matrix

para remuestrear un CountryStructure con las funciones definidas arriba.

Opcionalmente, si se desea modificar el comportamiento específico de cada función de remuestreo, se deben extender los siguientes métodos:

function (resamplefn::ResampleFunction)(cs::CountryStructure, rng = Random.GLOBAL_RNG)

function (resamplefn::ResampleFunction)(base::VarCPIBase, rng = Random.GLOBAL_RNG)

InflationEvalTools.ResampleFunction — Type

function (resamplefn::ResampleFunction)(cs::CountryStructure, rng = Random.GLOBAL_RNG)

Define el comportamiento general de función de remuestreo sobre CountryStructure. Se remuestrea cada una de las bases del campo base utilizando el método para objetos VarCPIBase y se devuelve un nuevo CountryStructure.

InflationEvalTools.ResampleFunction — Type

function (resamplefn::ResampleFunction)(base::VarCPIBase, rng = Random.GLOBAL_RNG)

Define el comportamiento general de función de remuestreo sobre VarCPIBase. Este método requiere una implementación específica del método sobre el par (AbstractMatrix, rng). Considera que el método de remuestreo podría extender los períodos de la serie de tiempo y ajusta las fechas apropiadamente.

InflationEvalTools.ResampleScrambleVarMonths — Type

ResampleScrambleVarMonths

Alias para InflationEvalTools.ScrambleVarMonths

InflationEvalTools.ResampleScrambleVarMonths — Type

scramblevar(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG)

Define cómo remuestrear matrices con las series de tiempo en las columnas. Utiliza la función interna scramblevar.

InflationEvalTools.ScrambleVarMonths — Type

struct ScrambleVarMonths <: ResampleFunction end

Define una función de remuestreo para remuestrear las series de tiempo por los mismos meses de ocurrencia. El muestreo se realiza de manera independiente para serie de tiempo en las columnas de una matriz.

InflationEvalTools.apply_trend — Method

apply_trend(cs::CountryStructure, trend::AbstractVector)

Función de aplicación de tendencia sobre las bases VarCPIBase del CountryStructure. El vector trend debe ser por lo menos del largo de periods(cs).

InflationEvalTools.get_param_function — Method

get_param_function(::ResampleFunction)

Función para obtener de una función de remuestreo la función que permite obtener las variaciones intermensuales promedio (o paramétricas), que finalmente sirven para construir la trayectoria paramétrica de inflación de la metodología de remuestreo.

InflationEvalTools.method_name — Method

method_name(resamplefn::ResampleFunction)

Función para obtener el nombre del método de remuestreo.

InflationEvalTools.param_gsbb_mod — Method

param_gsbb_mod(cs::CountryStructure)

Obtiene un CountryStructure paramétrico.

InflationEvalTools.param_gsbb_mod — Method

param_gsbb_mod(base::VarCPIBase)

Obtiene la matriz de variaciones intermensuales paramétricas para la metodología de remuestreo de Generalized Seasonal Block Bootstrap modificada que extiende las observaciones a 300 períodos. Devuelve una base de tipo VarCPIBase con las variaciones intermensuales paramétricas. Actualmente funciona solamente si base tiene 120 observaciones.

InflationEvalTools.param_sbb — Method

param_sbb(cs::CountryStructure)

Obtiene un CountryStructure paramétrico. Véase también param_sbb.

InflationEvalTools.param_sbb — Method

param_sbb(base::VarCPIBase)

Obtiene la matriz de variaciones intermensuales paramétricas para la metodología de remuestreo de Stationary Block Bootstrap. Devuelve una base de tipo VarCPIBase con las variaciones intermensuales promedio de los mismos meses de ocurrencia (también llamadas variaciones intermensuales paramétricas).

Esta definición también aplica a otras metodologías que utilicen como variaciones intermensuales paramétricas los promedios en los mismos meses de ocurrencia.

InflationEvalTools.scramblevar — Function

scramblevar(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG)

Copy and scramble every column of matrix vmat by months.

InflationEvalTools.scramblevar! — Function

scramblevar!(vmat::AbstractMatrix, rng = Random.GLOBAL_RNG)

Samples in-place every column of matrix vmat by months.