InflationEvalTools

InflationEvalTools.InflationEvalTools

InflationEvalTools.DEFAULT_SEED

InflationEvalTools.LOCAL_RNG

InflationEvalTools.RWTREND

InflationEvalTools.AbstractConfig

InflationEvalTools.AbstractInflationParameter

InflationEvalTools.ArrayTrendFunction

InflationEvalTools.CrossEvalConfig

InflationEvalTools.InflationParameter

InflationEvalTools.ResampleFunction

InflationEvalTools.ResampleFunction

InflationEvalTools.ResampleFunction

InflationEvalTools.ResampleScrambleVarMonths

InflationEvalTools.SimConfig

InflationEvalTools.TrendAnalytical

InflationEvalTools.TrendExponential

InflationEvalTools.TrendFunction

InflationEvalTools.TrendIdentity

InflationEvalTools.TrendRandomWalk

InflationEvalTools.DEFAULT_RESAMPLE_FN

InflationEvalTools.DEFAULT_TREND_FN

InflationEvalTools.ParamTotalCPI

InflationEvalTools.ParamTotalCPILegacyRebase

InflationEvalTools.ParamTotalCPIRebase

InflationEvalTools.ParamWeightedMean

InflationEvalTools.dict_config

InflationEvalTools.eval_metrics

InflationEvalTools.evalsim

InflationEvalTools.gentrayinfl

InflationEvalTools.get_param_function

InflationEvalTools.get_ranges

InflationEvalTools.makesim

InflationEvalTools.method_name

InflationEvalTools.method_name

InflationEvalTools.method_tag

InflationEvalTools.method_tag

InflationEvalTools.param_gsbb_mod

InflationEvalTools.param_gsbb_mod

InflationEvalTools.param_sbb

InflationEvalTools.param_sbb

InflationEvalTools.pargentrayinfl

InflationEvalTools.run_batch

InflationEvalTools.InflationEvalTools — Module

InflationEvalTools

Tipos, funciones y demás utilidades de simulación para evaluación de medidas inflación.

InflationEvalTools.DEFAULT_SEED — Constant

const DEFAULT_SEED

Semilla por defecto utilizada para el proceso de simulación y la reproducibilidad de los resultados.

InflationEvalTools.LOCAL_RNG — Constant

const LOCAL_RNG = StableRNG(0)

Esta constante se utiliza para fijar el generador de números aleatorios en cada proceso local, utilizando el generador StableRNG con semilla inicial cero. La semilla será alterada en cada iteración del proceso de simulación. Esto garantiza la reproducibilidad de los resultados por realización de remuestreo escogiendo la constante DEFAULT_SEED.

InflationEvalTools.RWTREND — Constant

RWTREND

Trayectoria de caminata aleatoria precalibrada para 240 períodos.

InflationEvalTools.AbstractConfig — Type

abstract type AbstractConfig{F <: InflationFunction, R <:ResampleFunction, T <:TrendFunction} end

AbstractConfig es un tipo abstracto para representar variantes de simulación que utilizan, en general, una función de inflación InflationFunction, una función de remuestreo ResampleFunction y una función de Tendencia TrendFunction. Contiene el esquema general de la simulación.

InflationEvalTools.AbstractInflationParameter — Type

Tipo abstracto para representar los parámetros de inflación

InflationEvalTools.ArrayTrendFunction — Type

abstract type ArrayTrendFunction <: TrendFunction end

Tipo para función de tendencia que almacena el vector de valores a aplicar a las variaciones intermensuales.

Utilización

function (trendfn::ArrayTrendFunction)(base::VarCPIBase{T}, range::UnitRange) where T

Especifica cómo aplicar la función de tendencia sobre un VarCPIBase con el rango de índices range.

InflationEvalTools.CrossEvalConfig — Type

CrossEvalConfig{F, R, T} <:AbstractConfig{F, R, T}

CrossEvalConfig es un tipo concreto que contiene una configuración base para generar simulaciones utilizando una muestra de los datos como set de entrenamiento y un período de n meses como período de evaluación . Recibe una función de inflación InflationFunction, una función de remuestreo ResampleFunction, una función de Tendencia TrendFunction, la cantidad de simulaciones deseadas [nsim], el último mes del set de entrenamiento [train_date] y el tamaño del período de evaluación en meses [eval_size].

Ejemplo

Considerando las mismas funciones de inflación, remuestreo, tendencia e inflación de evaluación:

percEq = InflationPercentileEq(80)
resamplefn = ResampleSBB(36)
trendfn = TrendRandomWalk()
paramfn = InflationTotalRebaseCPI(60)

Generamos una configuración del tipo CrossEvalConfig con 1000 simulaciones y utilizando el fin de set de entrenamiento hasta diciembre de 2012 y 24 meses de evaluación.

julia> config = CrossEvalConfig(percEq, resamplefn, trendfn, paramfn, 1000, Date(2012, 12), 24)
|─> Función de inflación     : PercEq-80.0
|─> Función de remuestreo    : ResampleSBB-36
|─> Función de tendencia     : TrendRandomWalk
|─> Inflación paramétrica    : InflationTotalRebaseCPI
|─> Simulaciones             : 1000
|─> Fin set de entrenamiento : 2012-12-01
|─> Meses de evaluación      : 24

InflationEvalTools.InflationParameter — Type

Tipo concreto para representar un parámetro de inflación computado con la función de inflación inflfn, el método de remuestreo resamplefn y función de tendencia trendfn.

Ver también: ParamTotalCPIRebase, ParamTotalCPI, ParamWeightedMean

InflationEvalTools.ResampleFunction — Type

abstract type ResampleFunction <: Function end

Tipo abstracto para funciones de remuestreo. Cada función debe extender como mínimo el método

resamplefn(vmat::AbstractMatrix, rng)::Matrix

para remuestrear un CountryStructure con las funciones definidas arriba.

Opcionalmente, si se desea modificar el comportamiento específico de cada función de remuestreo, se deben extender los siguientes métodos:

function (resamplefn::ResampleFunction)(cs::CountryStructure, rng = Random.GLOBAL_RNG)

function (resamplefn::ResampleFunction)(base::VarCPIBase, rng = Random.GLOBAL_RNG)

InflationEvalTools.ResampleFunction — Type

function (resamplefn::ResampleFunction)(cs::CountryStructure, rng = Random.GLOBAL_RNG)

Define el comportamiento general de función de remuestreo sobre CountryStructure. Se remuestrea cada una de las bases del campo base utilizando el método para objetos VarCPIBase y se devuelve un nuevo CountryStructure.

InflationEvalTools.ResampleFunction — Type

function (resamplefn::ResampleFunction)(base::VarCPIBase, rng = Random.GLOBAL_RNG)

Define el comportamiento general de función de remuestreo sobre VarCPIBase. Este método requiere una implementación específica del método sobre el par (AbstractMatrix, rng). Considera que el método de remuestreo podría extender los períodos de la serie de tiempo y ajusta las fechas apropiadamente.

InflationEvalTools.ResampleScrambleVarMonths — Type

ResampleScrambleVarMonths <: ResampleFunction

Define una función de remuestreo para remuestrear las series de tiempo por los mismos meses de ocurrencia. El muestreo se realiza de manera independiente para serie de tiempo en las columnas de una matriz.

InflationEvalTools.SimConfig — Type

SimConfig{F, R, T} <:AbstractConfig{F, R, T}

Tipo concreto que contiene una configuración base para generar simulaciones utilizando todos los datos como set de entrenamiento. Recibe una función de inflación InflationFunction, una función de remuestreo ResampleFunction, una función de Tendencia TrendFunction, una función de inflación de evaluación [paramfn] y la cantidad de simulaciones deseadas [nsim].

Ejemplo

Considerando las siguientes instancias de funciones de inflación, remuestreo, tendencia e inflación de evaluación:

percEq = InflationPercentileEq(80)
resamplefn = ResampleSBB(36)
trendfn = TrendRandomWalk()
paramfn = InflationWeightedMean()

Generamos una configuración del tipo SimConfig con 1000 simulaciones:

julia> config = SimConfig(percEq, resamplefn, trendfn, paramfn, 1000)
|─> Función de inflación     : PercEq-80.0
|─> Función de remuestreo    : ResampleSBB-36
|─> Función de tendencia     : TrendRandomWalk
|─> Inflación paramétrica    : InflationTotalRebaseCPI
|─> Simulaciones             : 1000

InflationEvalTools.TrendAnalytical — Type

TrendAnalytical{T} <: ArrayTrendFunction

Tipo para representar una función de tendencia definida por una función anónima. Recibe los datos de un CountryStructure o un rango de índices para precomputar el vector de tendencia utlizando una función anónima.

Ejemplos:

Para crear una función de tendencia a partir de una función anónima:

trendfn = TrendAnalytical(param_data, t -> 1 + sin(2π*t/12), "Tendencia sinusoidal")

o bien:

trendfn = TrendAnalytical(1:periods(param_data), t -> 1 + sin(2π*t/12), "Tendencia sinusoidal")

InflationEvalTools.TrendExponential — Type

TrendExponential{T} <: ArrayTrendFunction

Tipo concreto para representar una función de tendencia de crecimiento exponencial.

Constructores

function TrendExponential(cs::CountryStructure, rate::Real = 0.02f0)
function TrendExponential(range::UnitRange, rate::Real = 0.02f0)

Ejemplos:

# Crear una función de tendencia con crecimiento exponencial al 2% anual
trendfn = TrendExponential(gtdata, 0.02)

Utilización

function (trendfn::TrendExponential)(cs::CountryStructure)

Aplicación de tendencia TrendExponential sobre los objetos VarCPIBase que componen el CountryStructure.

trendfn = TrendExponential(gtdata, 0.02) 
trended_cs = trendfn(gtdata) 

InflationEvalTools.TrendFunction — Type

abstract type TrendFunction <: Function end

Tipo abstracto para manejar las funciones de tendencia.

Utilización

function (trendfn::TrendFunction)(cs::CountryStructure)

Aplica la función de tendencia sobre un CountryStructure y devuelve un nuevo CountryStructure.

InflationEvalTools.TrendIdentity — Type

TrendIdentity <: TrendFunction

Tipo concreto para representar una función de tendencia neutra. Es decir, esta función de tendencia mantiene los datos sin alteración.

Ejemplos:

# Crear una función de tendencia neutra. 
trendfn = TrendIdentity()

Utilización

function (trendfn::TrendIdentity)(cs::CountryStructure)

Aplicación de tendencia TrendIdentity sobre VarCPIBase. Se redefine este método para dejar invariante la base VarCPIBase.

trendfn = TrendIdentity() 
trended_cs = trendfn(gtdata) 

InflationEvalTools.TrendRandomWalk — Type

TrendRandomWalk{T} <: ArrayTrendFunction

Tipo para representar una función de tendencia de caminata aleatoria. Utiliza el vector de caminata aleatoria precalibrado en RWTREND.

Ejemplo:

# Crear la función de tendencia de caminata aleatoria
trendfn = TrendRandomWalk()

InflationEvalTools.DEFAULT_RESAMPLE_FN — Function

DEFAULT_RESAMPLE_FN

Define la funcón de remuestreo a utilizar por defecto en el ejercicio de simulación.

InflationEvalTools.DEFAULT_TREND_FN — Function

DEFAULT_TREND_FN

Define la funcón de tendencia a utilizar por defecto en el ejercicio de simulación.

InflationEvalTools.ParamTotalCPI — Method

ParamTotalCPI()

Función de ayuda para obtener la configuración del parámetro de inflación dado por la función de inflación del IPC, y el método de remuestreo y función de tendencia por defecto.

InflationEvalTools.ParamTotalCPILegacyRebase — Method

ParamTotalCPILegacyRebase()

Función de ayuda para obtener la configuración del parámetro de inflación dado por la función de inflación del IPC con cambio de base sintético, y el método de remuestreo y función de tendencia por defecto.

InflationEvalTools.ParamTotalCPIRebase — Method

ParamTotalCPIRebase()

Función de ayuda para obtener la configuración del parámetro de inflación dado por la función de inflación del IPC con cambio de base sintético, y el método de remuestreo y función de tendencia por defecto.

InflationEvalTools.ParamWeightedMean — Method

ParamWeightedMean()

Función de ayuda para obtener la configuración del parámetro de inflación dado por la media ponderada interanual y el método de remuestreo por defecto.

InflationEvalTools.dict_config — Method

dict_config(params::Dict)

Función para convertir diccionario a AbstractConfig.

InflationEvalTools.eval_metrics — Method

eval_metrics(tray_infl, tray_infl_pob)

Función para obtener diccionario con estadísticos de evaluación.

InflationEvalTools.evalsim — Method

evalsim(data_eval::CountryStructure, config::SimConfig; 
    rndseed = DEFAULT_SEED) -> (metrics, tray_infl)

Esta función genera la trayectoria paramétrica , las trayectorias de simulación y las métricas de evaluación utilizando la configuración SimConfig.

Genera el error cuadrático medio mse, el error estándar de simulación std_sim_error, la raíz del error cuadrático medio rmse, el error medio me y el error absoluto medio mae, así como un array con las trayectorias de simulación tray_infl.

Utilización

La función evalsim recibe un CountryStructure y un AbstractConfig del tipo SimConfig.

Ejemplo

Teniendo una configuración del tipo SimConfig y un set de datos gtdata_eval

julia> config = SimConfig(InflationPercentileEq(69), ResampleScrambleVarMonths(), TrendRandomWalk(), InflationTotalRebaseCPI(36, 2), 10_000)
SimConfig{InflationPercentileEq, ResampleScrambleVarMonths, TrendRandomWalk{Float32}}
|─> Función de inflación            : Percentil equiponderado 69.0
|─> Función de remuestreo           : Bootstrap IID por meses de ocurrencia
|─> Función de tendencia            : Tendencia de caminata aleatoria
|─> Método de inflación paramétrica : Variación interanual IPC con cambios de base sintéticos (36, 2)
|─> Número de simulaciones          : 10000

julia> results, tray_infl = evalsim(gtdata_eval, config)
┌ Info: Evaluación de medida de inflación
│   medida = "Percentil equiponderado 69.0"
│   remuestreo = "Bootstrap IID por meses de ocurrencia"
│   tendencia = "Tendencia de caminata aleatoria"
│   evaluación = "Variación interanual IPC con cambios de base sintéticos (36, 2)"
└   simulaciones = 10000
┌ Info: Métricas de evaluación:
│   mse = 2.307275f0
│   std_sim_error = 0.026808726787567138
│   rmse = 1.3116561f0
│   me = -1.3114622f0
│   mae = 1.3116561f0
└   corr = 0.97376233f0

InflationEvalTools.gentrayinfl — Method

gentrayinfl(inflfn::F, resamplefn::R, trendfn::T, csdata::CountryStructure; 
    K = 100, 
    rndseed = DEFAULT_SEED, 
    showprogress = true)

Computa K trayectorias de inflación utilizando la función de inflación inflfn::InflationFunction, la función de remuestreo resamplefn::TrendFunction y la función de tendencia trendfn::TrendFunction especificada. Se utilizan los datos en el CountryStructure dado en csdata.

A diferencia de la función pargentrayinfl, esta función no realiza el cómputo de forma distribuida.

Para lograr la reproducibilidad entre diferentes corridas de la función, y de esta forma, generar trayectorias de inflación con diferentes metodologías utilizando los mismos remuestreos, se fija la semilla de generación de acuerdo con el número de iteración en la simulación. Para controlar el inicio de la generación de trayectorias se utiliza como parámetro de desplazamiento el valor rndseed, cuyo valor por defecto es la semilla DEFAULT_SEED.

InflationEvalTools.get_param_function — Method

get_param_function(::ResampleFunction)

Cada función de remuestreo debe realizar la implementación de una función para obtener un CountryStructure con objetos VarCPIBase que contengan las variaciones intermensuales promedio (o paramétricas) que permitan construir la trayectoria paramétrica de inflación de acuerdo con el método de remuestreo dado en ResampleFunction.

Esta función devuelve la función de obtención de datos paramétricos.

Ejemplo

# Obtener la función de remuestreo 
resamplefn = ResampleSBB(36)
...

# Obtener su función para obtener los datos paramétricos 
paramdatafn = get_param_function(resamplefn)
# Obtener CountryStructure de datos paramétricos 
paramdata = paramdatafn(gtdata)

Ver también: param_sbb

InflationEvalTools.get_ranges — Method

get_ranges(cs::CountryStructure)

Función de ayuda para obtener tupla de rangos de índices para hacer slicing de los vectores de tendencia.

InflationEvalTools.makesim — Method

makesim(data, config::AbstractConfig; 
    rndseed = DEFAULT_SEED) -> (metrics, tray_infl)

Utilización

Esta función utiliza la función evalsim para generar un set de simulaciones en base a un CountryStructure y un AbstractConfig, y genera un diccionario results con todas las métricas de evaluación y con la información del AbstractConfig utilizado para generarlas. Adicionalmente genera un objeto con las trayectorias de inflación.

Ejemplos

makesim recibe un CountryStructure y un AbstractConfig, para trasladarlo a evalsim y generar las simulaciones. Almacena las métricas y los parámetros de simulación en el diccionario results, y Adicionalmente devuelve las trayectoria de simulacion.

julia> results, tray_infl = makesim(gtdata_eval, configA);
┌ Info: Evaluación de medida de inflación
│   medida = "Variación interanual IPC"
│   remuestreo = "Block bootstrap estacionario con bloque esperado 36"
│   tendencia = "Tendencia de caminata aleatoria"
└   simulaciones = 1000

┌ Info: Métricas de evaluación:
│   mse = 7.518966f0
│   std_sim_error = 0.48772313050091165
│   rmse = 1.9927315f0
│   me = 0.42103088f0
└   mae = 1.9927315f0

Exploramos el diccionario results:

julia> results
Dict{Symbol, Any} with 11 entries:
  :trendfn       => TrendRandomWalk{Float32}(Float32[0.953769, 0.948405, 0.926209, 0.902285, 0.832036, 0.825772, 0.799508, 0.789099, 0.764708, 0.757526  …  1.04656, 1.0…  :params        => (nothing,)
  :measure       => "Variación interanual IPC"
  :resamplefn    => ResampleSBB(36, Distributions.Geometric{Float64}(p=0.0277778))
  :me            => 0.421031
  :mae           => 1.99273
  :nsim          => 1000
  :rmse          => 1.99273
  :inflfn        => InflationTotalCPI()
  :mse           => 7.51897
  :std_sim_error => 0.487723

InflationEvalTools.method_name — Method

method_name(resamplefn::ResampleFunction)

Función para obtener el nombre del método de remuestreo.

InflationEvalTools.method_name — Method

method_name(resamplefn::TrendFunction)

Función para obtener el nombre de la función de tendencia.

InflationEvalTools.method_tag — Method

method_tag(resamplefn::ResampleFunction)

Función para obtener una etiqueta del método de remuestreo.

InflationEvalTools.method_tag — Method

method_tag(trendfn::TrendFunction)

Función para obtener una etiqueta de la función de tendencia.

InflationEvalTools.param_gsbb_mod — Method

param_gsbb_mod(cs::CountryStructure)

Obtiene un CountryStructure paramétrico.

InflationEvalTools.param_gsbb_mod — Method

param_gsbb_mod(base::VarCPIBase)

Obtiene la matriz de variaciones intermensuales paramétricas para la metodología de remuestreo de Generalized Seasonal Block Bootstrap modificada que extiende las observaciones a 300 períodos. Devuelve una base de tipo VarCPIBase con las variaciones intermensuales paramétricas. Actualmente funciona solamente si base tiene 120 observaciones.

InflationEvalTools.param_sbb — Method

param_sbb(cs::CountryStructure)

Obtiene un CountryStructure paramétrico. Véase también param_sbb.

InflationEvalTools.param_sbb — Method

param_sbb(base::VarCPIBase)

Obtiene la matriz de variaciones intermensuales paramétricas para la metodología de remuestreo de Stationary Block Bootstrap. Devuelve una base de tipo VarCPIBase con las variaciones intermensuales promedio de los mismos meses de ocurrencia (también llamadas variaciones intermensuales paramétricas).

Esta definición también aplica a otras metodologías que utilicen como variaciones intermensuales paramétricas los promedios en los mismos meses de ocurrencia.

InflationEvalTools.pargentrayinfl — Method

pargentrayinfl(inflfn::F, resamplefn::R, trendfn::T, csdata::CountryStructure; 
    K = 100, 
    rndseed = DEFAULT_SEED, 
    showprogress = true)

Computa K trayectorias de inflación utilizando la función de inflación inflfn::InflationFunction, la función de remuestreo resamplefn::TrendFunction y la función de tendencia trendfn::TrendFunction especificada. Se utilizan los datos en el CountryStructure dado en csdata.

A diferencia de la función gentrayinfl, esta función implementa el cómputo distribuido en procesos utilizando @distributed. Esto requiere que el paquete haya sido cargado en todos los procesos de cómputo. Por ejemplo:

using Distributed
addprocs(4, exeflags="--project")
@everywhere using HEMI 

Para lograr la reproducibilidad entre diferentes corridas de la función, y de esta forma, generar trayectorias de inflación con diferentes metodologías utilizando los mismos remuestreos, se fija la semilla de generación de acuerdo con el número de iteración en la simulación. Para controlar el inicio de la generación de trayectorias se utiliza como parámetro de desplazamiento el valor rndseed, cuyo valor por defecto es la semilla DEFAULT_SEED.

InflationEvalTools.run_batch — Method

run_batch(data, dict_list_params, savepath; 
    savetrajectories = true, 
    rndseed = DEFAULT_SEED)

La función run_batch genera paquetes de simulaciones con base en el diccionario de parámetros de configuración.

Utilización

La función recibe un CountryStructure, un diccionario con vectores que contienen parámetros de simulación y un directorio para almacenar archivos con las métricas de cada una de las evaluaciones generadas.

Ejemplo

Generamos un diccionario con parámetros de configuración para percentiles equiponderados, desde el percentil 60 hasta el percentil 80. Esto genera un diccionario con 21 configuraciones distintas para evaluación.

dict_prueba = Dict(
    :inflfn => InflationPercentileWeighted.(50:80), 
    :resamplefn => resamplefn, 
    :trendfn => trendfn,
    :paramfn => paramfn, 
    :nsim => 1000) |> dict_list`

Una vez creado dict_prueba, podemos generar el paquete de simulación utilizando run_batch.

julia> run_batch(gtdata_eval, dict_prueba, savepath)
...

Una vez generadas todas las simulaciones podemos obtener los datos mediante la función collect_results. Esta función lee los resultados desde savepath y los presenta en un DataFrame.

julia> df = collect_results(savepath); 
[ Info: Scanning folder `savepath` for result files.
[ Info: Added 31 entries.

julia> select(df, :measure, :mse)
31×2 DataFrame
 Row │ measure                   mse       
     │ String?                   Float32?  
─────┼─────────────────────────────────────
   1 │ Percentil ponderado 50.0  17.9208
   2 │ Percentil ponderado 51.0  16.8813
   3 │ Percentil ponderado 52.0  15.874
   4 │ Percentil ponderado 53.0  14.876
  ⋮  │            ⋮                  ⋮
  28 │ Percentil ponderado 77.0   2.85501
  29 │ Percentil ponderado 78.0   4.32532
  30 │ Percentil ponderado 79.0   6.33717
  31 │ Percentil ponderado 80.0   9.02473
                            23 rows omitted